requirements.txt
httpx>=0.25.0
pydantic>=2.0
python-dotenv

fastapi
uvicorn

pytest
pytest-mock

black>=24.3.0
isort>=5.13.2
